backend/requirements.txt
# ===== FastAPI backend =====
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic>=2.9.0

# ===== Environment =====
python-dotenv==1.0.1

# ===== PDF parsing =====
PyMuPDF>=1.25.0

# ===== Embeddings =====
sentence-transformers==2.7.0
torch>=2.2.0

# ===== Vector search =====
faiss-cpu>=1.11.0

# ===== LLM Providers =====
anthropic==0.34.2
google-genai==0.3.0

# ===== Utilities =====
numpy==1.26.4
scikit-learn==1.5.0
anthropic
chromadb